+++ PolarionMcpTools/GlobalUsings.cs
@@ -1,5 +1,6 @@
 
 global using System.Text;
+global using System.Text.RegularExpressions;
 global using System.Diagnostics.CodeAnalysis;
 global using System.ComponentModel;
 global using Polarion.Generated.Tracker;